+++ src/vs/platform/accessibility/browser/accessibleViewRegistry.ts
@@ -1,0 +1,37 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import { IDisposable } from 'vs/base/common/lifecycle';
+import { ContextKeyExpression } from 'vs/platform/contextkey/common/contextkey';
+import { ServicesAccessor } from 'vs/platform/instantiation/common/instantiation';
+
+export type CommandImplementation = (accessor: ServicesAccessor, args: unknown) => boolean | Promise<void>;
+
+export interface AccessibleViewImplentation {
+	priority: number;
+	name: string;
+	implementation: CommandImplementation;
+	when?: ContextKeyExpression | undefined;
+}
+
+export const AccessibleViewRegistry = new class AccessibleViewRegistry {
+	_implementations: AccessibleViewImplentation[] = [];
+
+	registerImplementation(implementation: AccessibleViewImplentation): IDisposable {
+		this._implementations.push(implementation);
+		return {
+			dispose: () => {
+				const idx = this._implementations.indexOf(implementation);
+				if (idx !== -1) {
+					this._implementations.splice(idx, 1);
+				}
+			}
+		};
+	}
+
+	getImplementations(): AccessibleViewImplentation[] {
+		return this._implementations;
+	}
+};
